+++ include/stcom.h
@@ -5,7 +5,7 @@ typedef struct stcom {
   int dummy;              /* just a dummy */
   int ieffcal;            /* 0 for frontend cal off, 1 for cal on */
   int ieffpcal;           /* 0 for phase cal off, 1 for pcal on */
-  long ieffrx;            /* what receiver (now lo frequency)*/
+  int ieffrx;            /* what receiver (now lo frequency)*/
   int rx_bits;            /* these are the 16bits sent to the SBC ,
                              top bit is cal offon */
   int antenna_sentoff;    /* if =1, antenna has just got new source command or azel offset, 
@@ -19,7 +19,7 @@ typedef struct stcom {
 			      in normal VLBI mode but but 0 if on/off or fivpt)
 			    4=stable on source (ionsor=1)*/
   int ant_scanning;       /*from vaxtime, -ve if waiting, + if scanning (scanmsec)*/
-  long ant_rfcentre;      /* from vaxtime broadcast: centre freq reported back from antenna,
+  int ant_rfcentre;      /* from vaxtime broadcast: centre freq reported back from antenna,
 			     should agree with value in rxlist table , if not, try to resend..*/
   double fsynth1,fsynth2; /*front end synthesizer(ULO) and 2nd LO synth in MHz*/
   int ant_is_listening;   /*controlled and read by rxvt, =1 if antenna is receiving commands from
